subscribe to both reportComment and moneyRequest notifications
